공지사항 핀 설정 파일/폴더명 변경 (Admin -> Management)
