+++ src/components/Management/ManagementComponent.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types'; // Import PropTypes
 
+import AdminPage from 'layout/NoticePin/AdminPage';
 import BoardTitle from '../BoardWrite/BoardTitle';
 import BoardText from '../BoardWrite/BoardText';
 import ManagementButton from 'components/Management/Button';
@@ -8,25 +9,25 @@ import StartendDate from '../../components/Management/StartendDate';
 import Local from '../../components/Management/SelectLocal';
 import Classify from '../../components/Management/Classify';
 
-import SearchChallenger from "./SearchChallenger.jsx";
-import ChallengerType from "./ChallengerType";
-import ChallengerPosition from "./ChallengerPosition";
-import styled from "styled-components";
+import SearchChallenger from './SearchChallenger.jsx';
+import ChallengerType from './ChallengerType';
+import ChallengerPosition from './ChallengerPosition';
+import styled from 'styled-components';
 
 const ButtonContainer = styled.div`
-display: flex;
-width: 120vh;
-justify-content: flex-end;
-align-items: center;
-margin-left: 16px;
+  display: flex;
+  width: 120vh;
+  justify-content: flex-end;
+  align-items: center;
+  margin-left: 16px;
 `;
 
 const ButtonContainerChallenger = styled.div`
-display: flex;
-width: 120vh;
-justify-content: center;
-align-items: center;
-margin-left: 16px;
+  display: flex;
+  width: 120vh;
+  justify-content: center;
+  align-items: center;
+  margin-left: 16px;
 `;
 
 const TypeComponent = ({ buttonStates }) => {
@@ -49,19 +50,16 @@ const TypeComponent = ({ buttonStates }) => {
           </ButtonContainer>
         </>
       )}{' '}
-
-     {buttonStates.challengerButton && (
-     <>
-        <SearchChallenger />
-        <ChallengerPosition />
-        <ChallengerType />
-        <ButtonContainerChallenger>
-          <ManagementButton />
-        </ButtonContainerChallenger>
-
-     </>
-     )} {' '} 
-
+      {buttonStates.challengerButton && (
+        <>
+          <SearchChallenger />
+          <ChallengerPosition />
+          <ChallengerType />
+          <ButtonContainerChallenger>
+            <ManagementButton />
+          </ButtonContainerChallenger>
+        </>
+      )}{' '}
     </div>
   );
 };
