+++ src/pages/Management/NoticePin/AdminManagePage.jsx
@@ -1,10 +1,10 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
 
-import AdminTitle from 'components/Admin/AdminTitle';
+import AdminTitle from 'components/Management/NoticePin/AdminTitle';
 import ManagementType from 'components/Management/ManagementType';
 // import AdminTable from 'components/Admin/AdminTable';
-import NewAdminTable from 'components/Admin/NewAdminTable';
+import NewAdminTable from 'components/Management/NoticePin/NewAdminTable';
 
 const AdminManagementContainer = styled.div`
   display: flex;
